requirements.txt
matplotlib==3.5.1
mtcnn==0.1.1
opencv_python==4.5.5.64
Pillow==9.1.0
streamlit==1.8.1
tensorflow-cpu==2.8.0